snapshot: extend the ExtraOption struct
Move ExtraOption to dedicated file and extend it for more usage cases.
